requirements.txt
flask
inflect
pyocd
pyusb